+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed text function ignoring camera when font set
 - Aligned the Emscripten version with Pyodide
 - Fixed the input_text variable
+- Stopped using the once_cell crate
 
 ## 2.3.4